requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pillow>=10.4.0
tensorflow>=2.20.0
numpy>=2.0.0
opencv-python>=4.10.0
pydantic>=2.9.0
aiofiles>=24.1.0
protobuf>=5.29.0
matplotlib>=3.9.0
